# generated/api/project-weather/rest/model_forecast.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Weather API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 */

package modules

type Forecast struct {
	Id string `json:"id,omitempty"`

	CityId string `json:"cityId"`

	ForecastDate string `json:"forecastDate"`

	Temperature string `json:"temperature"`

	Condition string `json:"condition"`

	CreatedAt int64 `json:"createdAt,omitempty"`

	UpdatedAt int64 `json:"updatedAt,omitempty"`
}

// AssertForecastRequired checks if the required fields are not zero-ed
func AssertForecastRequired(obj Forecast) error {
	elements := map[string]interface{}{
		"cityId":       obj.CityId,
		"forecastDate": obj.ForecastDate,
		"temperature":  obj.Temperature,
		"condition":    obj.Condition,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertForecastConstraints checks if the values respects the defined constraints
func AssertForecastConstraints(obj Forecast) error {
	return nil
}