# generated/api/project-weather/rest/model_page.go
// Code generated by OpenAPI Generator (https://openapi-generator.tech); DO NOT EDIT.

/*
 * Weather API
 *
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * API version: 1.0.0
 */

package modules

type Page struct {
	Limit int64 `json:"limit"`

	Next int64 `json:"next,omitempty"`

	Last int64 `json:"last,omitempty"`

	Count int64 `json:"count"`

	Current int64 `json:"current,omitempty"`

	Previous int64 `json:"previous,omitempty"`
}

// AssertPageRequired checks if the required fields are not zero-ed
func AssertPageRequired(obj Page) error {
	elements := map[string]interface{}{
		"limit": obj.Limit,
		"count": obj.Count,
	}
	for name, el := range elements {
		if isZero := IsZeroValue(el); isZero {
			return &RequiredError{Field: name}
		}
	}

	return nil
}

// AssertPageConstraints checks if the values respects the defined constraints
func AssertPageConstraints(obj Page) error {
	return nil
}